--- lib/Widgets/account_creation_page.dart
@@ -1,202 +1,0 @@
-import 'package:firebase_auth/firebase_auth.dart';
-import 'package:flutter/material.dart';
-
-// import '../main.dart';
-import 'account_page.dart';
-
-// TODO add firebase, including email + password and Google login capabilities
-FirebaseAuth auth = FirebaseAuth.instance;
-TextEditingController fName = TextEditingController();
-TextEditingController lName = TextEditingController();
-TextEditingController email = TextEditingController();
-TextEditingController password = TextEditingController();
-class AccountCreation extends StatefulWidget {
-  const AccountCreation({Key? key}) : super(key: key);
-
-  @override
-  _State createState() => _State();
-}
-class _State extends State<AccountCreation>{
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        backgroundColor: Colors.indigo[900],
-        foregroundColor: Colors.white,
-        shadowColor: Colors.blueGrey,
-        shape: const RoundedRectangleBorder(
-            borderRadius: BorderRadius.vertical(
-              bottom: Radius.circular(35),
-            )
-        ),
-        title: const Text('Account Creation'),
-        centerTitle: true,
-      ),
-      body: Center(
-        child: SingleChildScrollView(
-          child: Column(
-            children: const <Widget>[
-              AccountForm(),
-            ])
-        ),
-
-
-
-
-
-
-      ),
-    );
-
-  }
-
-}
-
-class AccountForm extends StatefulWidget {
-  const AccountForm({Key? key}) : super(key: key);
-
-  @override
-  AccountState createState() => AccountState();
-}
-
-class AccountState extends State<AccountForm> {
-  final _formKey = GlobalKey<FormState>();
-
-
-  @override
-  Widget build(BuildContext context) {
-    return Form(
-        key: _formKey,
-        child: SingleChildScrollView(
-        child: Column(
-            crossAxisAlignment: CrossAxisAlignment.center,
-            children: <Widget>[
-              Padding(
-                padding: const EdgeInsets.symmetric(
-                    horizontal: 8, vertical: 16),
-                child: TextFormField(
-                  controller: fName,
-                  decoration: const InputDecoration(
-                    border: UnderlineInputBorder(),
-                    labelText: 'First Name',
-                    hintText: 'Enter your First Name',
-                  ),
-
-                  validator: (value) {
-                    if (value == null || value.isEmpty) {
-                      return 'Please enter your first name.';
-                    }
-                    return null;
-                  },
-                ),
-              ),
-              Padding(
-                padding: const EdgeInsets.symmetric(
-                    horizontal: 8, vertical: 16),
-                child: TextFormField(
-                  controller: lName,
-                  decoration: const InputDecoration(
-                    border: UnderlineInputBorder(),
-                    labelText: 'Last Name',
-                    hintText: 'Enter your Last Name',
-                  ),
-
-                  validator: (value) {
-                    if (value == null || value.isEmpty) {
-                      return 'Please enter your last name.';
-                    }
-                    return null;
-                  },
-                ),
-              ),
-              Padding(
-                padding: const EdgeInsets.symmetric(
-                    horizontal: 8, vertical: 16),
-                child: TextFormField(
-                  controller: email,
-                  decoration: const InputDecoration(
-                    border: UnderlineInputBorder(),
-                    labelText: 'Email',
-                    hintText: 'Enter your email address',
-
-                  ),
-
-                  validator: (value) {
-                    bool sym;
-                    if (value != null && value.contains('@')) {
-                      sym = true;
-                    }
-                    else {
-                      sym = false;
-                    }
-                    if (value == null || value.isEmpty || !sym) {
-                      return 'Please enter a valid email address.';
-                    }
-                    return null;
-                  },
-                ),
-              ),
-              Padding(
-                padding: const EdgeInsets.symmetric(
-                    horizontal: 8, vertical: 16),
-                child: TextFormField(
-                  controller: password,
-                  decoration: const InputDecoration(
-                    border: UnderlineInputBorder(),
-                    labelText: 'Password',
-                    hintText: 'Enter a password',
-                  ),
-
-                  validator: (value) {
-                    if (value == null || value.isEmpty || value.length < 8) {
-                      return 'Password must have at least 8 characters';
-                    }
-                    return null;
-                  },
-                ),
-              ),
-              Container(
-                height: 50,
-                width: 250,
-                decoration: BoxDecoration(
-                    color: Colors.blueGrey, borderRadius: BorderRadius.circular(20)),
-                child: FlatButton(
-                  onPressed: () {
-                    createUserWithEmailAndPassword();
-                    Navigator.pop(
-                        context, MaterialPageRoute(builder: (_) => AccountPage()));
-                  },
-                  child: Text(
-                    'Create Account',
-                    style: TextStyle(color: Colors.white, fontSize: 25),
-                  ),
-                ),
-              ),
-            ]
-        )
-    )
-    );
-  }
-}
-
-//TODO create account creation form, implement firebase logging and storing
-
-void createUserWithEmailAndPassword() async {
-   try {
-     UserCredential userCredential = await FirebaseAuth.instance
-         .createUserWithEmailAndPassword(
-         email: email.text,
-         password: password.text
-     );
-   } on FirebaseAuthException catch (e) {
-     if (e.code == 'weak-password') {
-       print('The password provided is too weak.');
-     } else if (e.code == 'email-already-in-use') {
-       print('The account already exists for that email.');
-     }
-   } catch (e) {
-     print(e);
-   }
-}
-
